requirements.txt
streamlit
yfinance
pandas
plotly
python-dotenv
